+++ core/src/main/java/com/graphhopper/GraphHopper.java
@@ -249,8 +249,8 @@ public class GraphHopper {
      * <pre>
      * {@code
      *   hopper.setProfiles(
-     *     new Profile("my_car").setVehicle("car").setWeighting("shortest"),
-     *     new Profile("your_bike").setVehicle("bike").setWeighting("fastest")
+     *     new CustomProfile("my_car").setVehicle("car"),
+     *     new CustomProfile("your_bike").setVehicle("bike")
      *   );
      *   hopper.getCHPreparationHandler().setCHProfiles(
      *     new CHProfile("my_car"),