+++ core/src/main/java/com/graphhopper/routing/weighting/AbstractWeighting.java
@@ -40,25 +40,14 @@ public abstract class AbstractWeighting implements Weighting {
         this.turnCostProvider = turnCostProvider;
     }
 
-    @Override
-    public boolean edgeHasNoAccess(EdgeIteratorState edgeState, boolean reverse) {
+    protected boolean edgeHasNoAccess(EdgeIteratorState edgeState, boolean reverse) {
+        if (edgeState.getBaseNode() == edgeState.getAdjNode())
+            throw new IllegalStateException("Unexpected loop-edge at node: " + edgeState.getBaseNode());
         return reverse ? !edgeState.getReverse(accessEnc) : !edgeState.get(accessEnc);
     }
 
-    /**
-     * In most cases subclasses should only override this method to change the edge-weight. The turn cost handling
-     * should normally be changed by passing another {@link TurnCostProvider} implementation to the constructor instead.
-     */
-    public abstract double calcEdgeWeight(EdgeIteratorState edgeState, boolean reverse);
-
     @Override
     public long calcEdgeMillis(EdgeIteratorState edgeState, boolean reverse) {
-        // special case for loop edges: since they do not have a meaningful direction we always need to read them in
-        // forward direction
-        if (edgeState.getBaseNode() == edgeState.getAdjNode()) {
-            reverse = false;
-        }
-
         if (reverse && !edgeState.getReverse(accessEnc) || !reverse && !edgeState.get(accessEnc))
             throw new IllegalStateException("Calculating time should not require to read speed from edge in wrong direction. " +
                     "(" + edgeState.getBaseNode() + " - " + edgeState.getAdjNode() + ") "
@@ -102,7 +91,7 @@ public abstract class AbstractWeighting implements Weighting {
 
     @Override
     public String toString() {
-        return getName() + "|" + speedEnc.getName().split("$")[0];
+        return getName() + "|" + speedEnc.getName();
     }
 
 }